+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## eol 0.7.5 (not yet released)
 
+- [Issue #3] Add "dos", "windows" and "unix" aliases for specifying an eol
+  style. E.g. can now do `eol -c dos foo.bat` to convert to DOS-style EOLs.
 - [Issue #4] Fix a few Python 3 compat issues, add some more test cases.
 
 ## eol 0.7.4